Start caching generated bitmap ImageAsset instances
Currently we're just caching Coil's RequestResult instance, which
contains a Drawable. That isn't something we can use in Compose, so we
need to convert that to Bitmap, and then to an ImageAsset. The current
code path is performing that conversion on every composition, which
obviously isn't great.

This comment fixes that by creating our result entities of the same
name, but contain an ImageAsset rather than a drawable, meaning we only
perform the conversion once.