Fix logging in MaterialThemeFromMdcTheme
Currently we use a reflective framework call to Theme.getKey()
to use a key for remembering the result. That doesn't work on
places where the SDK greylist is in operation, and we currently
log every time on those API levels.

This commit fixes that by going back to a traditional reflective
call setup, and removing all logging.