Move away from using `WithConstraints` (#30)
* Move away from using WithConstraints

WithConstraints is not suitable for how we're using
it, so we're moving to Modifier.onSizeChanged(). This
required some changes to how CoilImage is implemented.

While here we now use the new launchInComposition()
function.

* Utilize actor pattern

* Stop returning to composition to callback

* Add some doc for RequestActor

* Fix Coil request being launch on every composition

Due to using the wrong key parameter to the outer `remember`. Also
fixed the state which holds the current callback.

* Update to Coil v0.11.0

* Address feedback