Stop refetching on every size change by default
Currently we invoke Coil on every size change, which is correct
but a lot of wasted work most of the time. We no longer re-fetch
on every size change. If developers want to customize this, they
can use the new `shouldRefetchOnSizeChange` lambda parameter
to allow re-fetching as they see fit.