Fix using CoilImage with an implicit size
Currently CoilImage only works when you specific a size
Modifier, since it waits for onPositioned to be called,
and only starts a Coil request when it has a non-zero
size. If no Modifier is provided, the CoilImage will
have a zero size and we never execute a Coil request.

Fixed by moving back to WithConstraints to move the request
size logic to be based on the incming constraints, rather
than the final size, which is probably more correct anyway.