Fix GlideImage crashing when resumed
Glide seems to perform all sorts of implicit and automatic
behavior attached to the host lifecycle. To stop that, we
now use the application context as the source for our
default RequestManager.

Also updated our coroutine wrapper to swallow/no-op if Glide
calls our listener/target a second time.

Fixes #196