[Pager] Make currentPager clearer
Reading the selected page is a common use for pagers, especially when
used with things like Tabs. This PR clarifies what currentPage and
currentPageOffset to mean the 'selected page', and thus only get updated
after a scroll has completed.

This allows us to tidy up the Tab offset modifier and make it a
provided function in the library.

Internally, we've now separated the 'selected page' from the internal
'layout page' (and offset), which allows us to set the selected page
only once a scroll has finished.

Fixes #280