Fix recomposition affecting scrolling
If the Pager composable is recomposed during a scroll
it currently resets back to the original page. This is
because the recomposition will eventually update the
PagerState.pageCount count value, which currently triggers
a page layout update.

There's no real need to update the layout when this happens,
as the updated currentPage will do it as necessary.

Fixes #585