[Navigation] Add navigation module (#574)
* Add navigation module

New module for allowing custom transition in Navigation Compose.

New API: AnimatedNavHost

* Fix error in Navigation docs

Correct the docs to say ExperimentalAnimationApi instead of
ExperimentalPermissionApi.

Also simplified the sample in the README.

* Change module name to navigation-animation

Changing the module name from navigation to navigation-animation to be
more precise about what the module is actually for. Renamed all of the
packages and files to reflect this change.

* Add pop animation parameters

Added popEnter and popExit parameters to the animation APIs.

Also updated to latest SNAPSHOT for needed NavController changes.

* Tidy up libs.versions.toml

* Fix activity-compose transitive dependency

* Alias Compose 1.0.1 to the current version

* Remove unneeded test dependency

Navigation does not need to clear package data using the
testInstrumentationRunner.

* Reduce default animation time

Reducing the default animation time from 2000ms to 700ms to match the
material guidelines.

Also adjusted the animation time in the samples.

* Add rememberAnimatedNavController API

Instead of making developers manually add the AnimatedComposeNavigator
as part of their sample, we can just provide a
rememberAnimatedNavController API that does it for them.

* Add rememberAnimatedNavController API

Instead of making developers manually add the AnimatedComposeNavigator
as part of their sample, we can just provide a
rememberAnimatedNavController API that does it for them.

Co-authored-by: Jeremy Woods <jbwoods@google.com>
Co-authored-by: Chris Banes <chris@banes.me>