[Navigation Animation] Add AnimatedContentScope and AnimatedVisiblityScope (#662)
* Use `AnimatedContentScope` for Navigation Transitions

Rather than directly pass the `initial` and `target` parameters, use `AnimatedContentScope<NavBackStackEntry>` to allow direct access to the `initialState` and `targetState` and enable usage of `slideIntoContainer` and `slideOutOfContainer` style transitions.

* Add `AnimatedVisibilityScope` to composable content

Allow the `content` of a `composable` destination to use the `animateEnterExit` Modifier via the `AnimatedVisibilityScope`.

* Fix spotless errors on `AnimatedNavHostSample`

* Re-add `initial` and `target` parameters to transitions

As the internal `AnimatedContent` needs to use the `id` of each `NavBackStackEntry` as the `key` for saving state, we need to use `AnimatedContentScope<String>` rather than `AnimatedContentScope<NavBackStackEntry>`, which necessitates manually passing through the final `initial` and `target` entries.

Co-authored-by: Chris Banes <chris@banes.me>