Fix ANR when showing bottom sheets
Our `internalShow` method could end up in an infinite loop sometimes (yay, recursion!). We now use the official `show` API which *should* work. Unfortunately, our `show` calls all get cancelled but Swipeable still settles in the closest state. This seems to always be the expanded/half-expanded state. We look at the state's targetValue in our try's `finally` block to figure out if the state will settle in an expanded state and call the `onSheetShown` callback based on that.

Fixes #660