Reliably mark showing sheet's back stack entry as complete
Our `show` calls are getting cancelled due to a race condition (https://issuetracker.google.com/issues/200980998) so we can't rely on their completion as a signal that the sheet is shown and we can mark the sheet's back stack entry's transition as complete. Using `withFrameNanos`, we try to wait until the sheet is visible.

Test: Existing tests still pass (as they did despite the race condition...). Will try to create tests that fail before the fix. Manual testing apart from that.
Fixes: #725, #720, #700