[Pager] PagerState's lazyListState public
Make PagerState's lazyListState public.

I'd like to propose having the lazyListState of the PagerState to become public. A use case that I am needing the lazyListState is to check the `layoutInfo.visibleItemsInfo` of the lazyListState to check if an item is currently visible. I also saw the question as the comment, and my answer is Yes.