[Pager] Deprecate PagerState.targetPage
targetPage is deprecated in favor of currentPage as currentPage property is now being updated right after we over scrolled the half of the previous current page. If you still think that you need targetPage, not currentPage please file a bug as we are planning to remove this property in future.