Avoid crashing when the start destination changes
Since moving to using visibleEntries, if you change your start
destination dynamically, instead of dropping the graph and rebuilding,
the AnimatedNavHost crashes.

The reason for this crash is that setting the graph synchronously pops
the entire back stack, which clears all of the visibleEntries so there
is a frame in AnimatedNavHost where it fails to find any of the old
entries.

We should move setNavGraph over to using the popWithTransition function,
but there is pre-req works that needs to be done. So as a stop gap, we
will return null from visibleEntries and disable the animations on
AnimatedNavHost.

Fixes #1162