Alter use of READ_EXTERNAL_STORAGE permission from permissions samples.
fixes #1679.

RequestMultiplePermissionsSample was using READ_EXTERNAL_STORAGE, but
this permission was removed at API Level 33 for more detailed items.
It leads to the state where rememberMultiplePermissionsState() never
count it as approved whereas launchMultiplePermissionRequest() will
never ask the user for it as a valid permission, and the entire sample
never gets to the "both approved" state.

Use another permission entry which is still valid at any API level.