Ensure Navigation Material properly handles back for nested nav
If there is a nested NavGraph under the current bottomSheet destination
and you do a back press, instead of the bottomSheet which is the topmost
destination being popped, the nested NavHost underneath will be popped
instead. This is caused by bottomSheet not properly intercepting back presses.

This change added a BackHandler to ensure that bottomsheets correctly
respect the system back hierarchy for back events.

Fixes: #1726